替换cache策略为weakvalue

+++ common-ramcache/src/main/java/com/windforce/common/ramcache/service/CacheFinder.java
@@ -58,5 +58,5 @@ public interface CacheFinder<PK extends Comparable<PK> & Serializable, T extends
 	 * 
 	 * @return
 	 */
-	int getAllSize();
+	long getAllSize();
 }
